+++ backend/utils/errorController.ts
@@ -1,5 +1,5 @@
 import { CastError, Error as MongooseError } from 'mongoose';
-import AppError, { AppErrorType } from '../utils/appError';
+import AppError, { AppErrorType } from './appError';
 import { Request, Response, NextFunction } from 'express';
 import type { MongoServerError } from 'mongodb';
